Now select language container is new module (because it used by two headers). Minor fixes
