+++ SportHub/src/modules/SelectLanguage/index.js
@@ -1,0 +1,3 @@
+import SelectLanguage from "./components/SelectLanguage"
+
+export default SelectLanguage
