+++ SportHub/src/modules/SelectLanguage/styles/style.module.scss
@@ -1,0 +1,15 @@
+@import "src/styles/mixins";
+
+.select {
+  border: none;
+  outline: none;
+  font-weight: bolder;
+  font-family: sans-serif;
+  font-size: 14px;
+
+  &:hover {
+    cursor: pointer;
+    border: none;
+    outline: none;
+  }
+}
