--- SportHub/src/modules/UserHeader/style/style.module.scss
@@ -1,161 +1,0 @@
-@import "src/styles/mixins";
-
-select::-ms-expand {
-  display: none;
-}
-
-.header{
-  width: 100%;
-  display: flex;
-  align-items: center;
-  .searchbar{
-      display: flex;
-      left: 18vw;
-      top: 0;
-      @include flex-center;
-      background: var(white);
-      width: 25vw;
-      height: 12vh;
-      font-size: 18px;
-      color: rgb(189, 189, 189);
-      border-right: 1px solid rgb(240, 240, 240);
-    &:hover{
-      cursor: pointer;
-    }
-
-    .img_icon{
-      height: 18px;
-      width: 18px;
-    }
-  }
-  .media
-  {
-    display: flex;
-    left: 43vw;
-    top: 0;
-    @include flex-center;
-    background: var(white);
-    width: 25vw;
-    height: 10vh;
-    font-size: 15px;
-    color: rgb(189, 189, 189);
-    border-right: 1px solid rgb(240, 240, 240);
-    .share
-    {
-      &:hover{
-        cursor: pointer;
-      }
-    }
-    .icons{
-      padding-left: 10px;
-      width: 35px;
-      height: 35px;
-      &:hover{
-        cursor: pointer;
-      }
-    }
-
-  }
-.input
-{
-  padding-bottom: 20px;
-  width: 20vw;
-  left: 1vw;
-  outline: none;
-  border: none;
-  outline-color: white;
-
-  &:focus{
-    outline: none;
-    border: none;
-    outline-color: white;
-  }
-  &:hover{
-    outline: none;
-    border: none;
-    outline-color: white;
-  }
-
-}
-
-.btn_signup
-{
-  display: flex;
-  @include flex-center;
-  padding: 10px;
-  border: none;
-  margin-right: 2vw;
-  color: rgb(216, 38, 52);
-  font-family: sans-serif;
-  font-weight: bolder;
-  background-color: white;
-  font-size: 14px;
-
-
-  &:hover{
-    cursor: pointer;
-    color: white;
-    background-color: rgb(216, 38, 52);
-  }
-}
-.auth
-{
-   display: flex;
-  left: 68vw;
-  top: 0;
-  @include flex-center;
-  background: var(white);
-  width: 21vw;
-  height: 12vh;
-  font-size: 18px;
-  color: rgb(189, 189, 189);
-
-.profile_img
-{
-  cursor: pointer;
-  border-radius: 50%;
-  height: 40px;
-  width: 40px;
-}
-  .user_name_text{
-    cursor: pointer;
-    color: black;
-    font-size: 14px;
-    font-family: sans-serif;
-    font-weight: bolder;
-    padding-right: 3px;
-  }
-  .profile_img_icon{
-    cursor: pointer;
-    width: 10px;
-    height: 10px;
-  }
-}
-.languages
-{
-   display: flex;
-  left: 89vw;
-  top: 0;
-  @include flex-center;
-  background: var(white);
-  width: 11vw;
-  height: 12vh;
-  font-size: 18px;
-  color: rgb(189, 189, 189);
-
-  .slect{
-    border: none;
-    outline: none;
-    font-weight: bolder;
-    font-family: sans-serif;
-    font-size: 14px;
-
-    &:hover{
-      cursor: pointer;
-      border: none;
-      outline: none;
-    }
-  }
-}
-
-}
